08_MONITORING/GENAI/EVALS/requirement.txt
langchain_openai
langchain_core
python-dotenv
openevals